# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TRADE FAST GOLD Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /wallets/transfer:
    post:
      operationId: walletTransfer
      tags: [wallets]
      summary: Transfer between wallets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletTransferInput"
      responses:
        "200":
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResult"

  # ─── DEPOSITS ─────────────────────────────────────────────────────────────
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List user withdrawals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, declined]
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Request withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Created withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{id}:
    get:
      operationId: getWithdrawal
      tags: [withdrawals]
      summary: Get withdrawal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # ─── TRANSACTIONS ─────────────────────────────────────────────────────────
...
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Paginated withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWithdrawalPage"
  /admin/withdrawals/{id}/approve:
    patch:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve withdrawal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{id}/decline:
    patch:
      operationId: declineWithdrawal
      tags: [admin]
      summary: Decline withdrawal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Declined withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, firstName, lastName, role, kycStatus, referralCode, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        kycStatus:
          type: string
          enum: [not_submitted, pending, approved, rejected]
        referralCode:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        country:
          type: string
        avatarUrl:
          type: string

    # ─── DASHBOARD ──────────────────────────────────────────────────────────
...
    WalletTransferInput:
      type: object
      required: [fromWalletType, toWalletType, amount]
      properties:
        fromWalletType:
          type: string
          enum: [main, profit, bonus, referral]
        toWalletType:
          type: string
          enum: [main, profit, bonus, referral]
        amount:
          type: number
    TransferResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        fromBalance:
          type: number
        toBalance:
          type: number

    # ─── DEPOSITS ───────────────────────────────────────────────────────────
...
    Withdrawal:
      type: object
      required: [id, userId, amount, method, status, walletAddress, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [bank_account, crypto_wallet]
        status:
          type: string
          enum: [pending, approved, declined]
        walletAddress:
          type: string
        bankDetails:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        processedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    WithdrawalInput:
      type: object
      required: [amount, method, walletAddress]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank_account, crypto_wallet]
        walletAddress:
          type: string
        bankDetails:
          type: string

    # ─── TRANSACTIONS ───────────────────────────────────────────────────────
...
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ─── ADMIN ──────────────────────────────────────────────────────────────
...
    AdminWithdrawalPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...